Fix Supabase integration: ES256 JWT verification and schema drift
The backend rejected every logged-in request with "Signed JWT rejected:
Another algorithm expected" because NimbusJwtDecoder defaulted to RS256
while Supabase signs access tokens with ES256. Configure the decoder to
accept ES256 (and RS256 for future key rotation).

Reconcile the schema with the JPA entities so the backend boots under
ddl-auto=validate: budget is a double in the entity, so the column is
now double precision instead of numeric. (The live database and its
title column / update+delete RLS policies were reconciled directly.)

+++ supabase/schema.sql
@@ -57,7 +57,7 @@ create trigger on_auth_user_created
 create table if not exists public.investment_recommendations (
   id bigint generated always as identity primary key,
   user_id uuid not null references auth.users(id) on delete cascade,
-  budget numeric not null,
+  budget double precision not null,
   timeline text not null,
   risk_tolerance text not null,
   strategies jsonb not null,
